+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: rabouzia <rabouzia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 14:29:41 by rabouzia          #+#    #+#             */
-/*   Updated: 2023/11/14 17:49:42 by rabouzia         ###   ########.fr       */
+/*   Updated: 2023/11/17 16:09:00 by rabouzia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
